fix(vim): remove plugin jelera/vim-javascript-syntax
The plugin conflicts with pangloss/vim-javascript and mwx/vim-jsx.

Specifically, consider this example:

class Something extends Component {
  render() {
    return <SomethingElse {...this.props} />;
  }
}

When the cursor is positioned on the opening brace "{" in the
line with "render()", pressing % jumps to the closing brace after
this.props (!).

Furthermore, the syntax highlighting on the braces around this.props
is off.