fix(vim): update_vim.sh could change dirs too early
The EXIT trap was triggered when "otool ... | grep Rub" failed.

This is correct, but there was no indication as to what happened
except that the "make install" somehow did not work (because
"cd -" had been executed).

The "grep Rub" command was outdated anyway ("grep ruby" would work),
so replacing that and providing a descriptive error message.