fix(tmux): fix synchronize-panes binding
The binding was erroneously global, but should only work with a prefix.
(see tmux manpage: "-n is an alias for -T root", where the root table is
used for keys pressed without the prefix key)

(Using C-X s as a global key in particular conflicts with C-X in fzf.vim)

Furthermore, we can just toggle it by not specifying on/off.