Move all files except .vim and bin to dotfiles directory
The dotfiles are being migrated to yadm. In order to keep the previous
history, this happens in the original dotfiles repository. Created
a branch called "yadm" there, then checked out the repo with

    yadm clone -b yadm <dotfiles-repo-url>

We use yadm's default work-tree $HOME. Therefore to keep the current
directory structure, moving everything checked out there into the
dotfiles directory.

Exceptions: ".vim" and "bin" are (now) already at the correct location.

Note: ".gitignore" and ".gitmodules" were moved to dotfiles, too,
thereby losing their functionality for this repo. This is alright
because instead of .gitignore, yadm sets showUntrackedFiles,
and the .gitmodules will be re-initialized with yadm later on.