feat(bash): Source bashrc from bash_profile only on non-openSUSE distros
openSUSE parses $HOME/.bashrc from /etc/profile (!). In a login shell,
this would mean the following order of invocations:

  /etc/profile
  $HOME/.bashrc
  $HOME/.bash_profile
  $HOME/.bashrc

And thus it means
a) sourcing .bashrc before .bash_profile and
b) sourcing .bashrc twice, if it is sourced from .bash_profile

We only solve b) which is good enough for now.