fix(bash): Fix git completion on Macos
Outside tmux, the completion setup in bashrc, namely

    $(brew --prefix)/share/bash-completion/bash_completion

, negated the earlier completion setup in bash_profile.

Inside tmux, the bash_profile was not sourced at all, hence no
git completion was available either.

To fix, ensure the bash_profile setup is called *after* the bashrc invocations.
Do this in bashrc and not in bash_profile so that the completions
are also available in tmux where only bashrc is sourced.

(Remember, we only use bashrc in tmux because bash_profile is meant
for one-off settings which are inherited by tmux. For example, PATH
settings would be duplicated otherwise. (3f82de1))