feat(bin): Rotate 1st of month/year in db_backup.sh (if available)
Before, monthly backups would usually store the last day of the month:
2021-01-31
2021-02-28
2021-03-31

or the last day in which a backup happened, e.g. 2021-04-25.

Now the backups will be at the first of the month.
2021-01-01
2021-02-01
2021-03-01

or the last day in which a backup happened, e.g. 2021-03-25.

On the one hand, this makes it easier to scan through the backups.
On the other hand, this leads to a situation where it is possible
to have two backups from the same month.