fix(nvim): better color contrast in lualine diff component
The background of lualine_a is too bright for the diff component with the
gruvbox theme. It is apparently assumed that the diff component lives in
section B.

For example, when in section A, ":hi" shows for lualine_a_diff_added_normal and
lualine_a_diff_added_terminal

    guifg=#b8bb26 guibg=#a89984

, and for lualine_a_diff_modified_command

guifg=#8ec07c guibg=#b8bb26

According to https://webaim.org/resources/contrastchecker/, the latter has a
almost non-existing contrast of 1.01.