feat(nvim): fix git-conflict highlighting
The default for current is DiffText, which is a bright yellow.
Especially the label line (i.e. the top line) will get too bright and
difficult to read with that.

This happened with https://github.com/ellisonleao/gruvbox.nvim/commit/d3b86c23e0fbbcd338900241ca55f3cef1da6111

The gruvbox author recognizes as much on the readability of DiffText,
see
https://github.com/ellisonleao/gruvbox.nvim/discussions/239#discussioncomment-7154441