feat(mutt): call "notmuch new" when executing sync-mailbox
Szenario:
1. navigate to a gmail.newsletter-mailbox
2. Press "N" to mark an email read
3. navigate to the virtual mailbox "Unread newsletters"

Expected: The email just marked read is not shown
Actual: The email is shown

To update the virtual inbox, two steps are required:
1. the mail file must be renamed on disk. This happens when sync-mailbox
   is called, either explicitly with "$" or implicitly by navigating to
   a new mailbox
2. "notmuch new" must be executed

This macro still requires pressing "$" (I would actually need a
sync-mailbox hook, see https://github.com/neomutt/neomutt/issues/1026)

I tried to configure

    macro index N '<toggle-new><modify-labels>!unread<enter>'

but the "modify-labels" function is only available in a notmuch folder,
see https://github.com/neomutt/neomutt/issues/742#issuecomment-354751480

I then tried to get rid of the need to manually press "$" by
implementing a folder hook that is executed upon opening the virtual
mailbox, as suggested in
https://github.com/neomutt/neomutt/issues/1026#issuecomment-428801549
The documented hook (with "exec") did not work, but

    folder-hook notmuch:// 'push "<shell-escape>notmuch new --no-hooks<enter>"'

executed notmuch. However, even though the tag is updated by
the folder hook, neomutt does not reflect this in the just opened
virtual mailbox. Only after entering a different mailbox and re-entering
it do I see this change. The debug log shows that the hook is executed
asynchronously (test with "<shell-escape echo hello; sleep 5<enter>" and
look for the log's timestamps).

Note that it works the other way around:

1. Navigate to a virtual mailbox
2. Press "N" to mark an email read
3. Navigate to the real gmail-mailbox

Expected: The email just marked read is not shown
Actual: As expected.

Explanation: The filename changes as soon as I open the new mailbox (due
to sync-mailbox, as usual). Note that the mail does not immediately
vanish from the virtual mailbox after pressing "N" - which is as
expected since sync-mailbox has not been called yet. In this case, just
by switching to a new mailbox, apparently the notmuch-tag is updated,
too.

https://github.com/neomutt/neomutt/issues/1485