feat(mutt): synchronize notmuch tags with fswatch
Before starting neomutt, start notmuch-new-inotify.sh.

When in a normal neomutt mailbox (i.e. not a notmuch virtual mailbox),
and when marking the message read, then once the mailbox
is synchronized (either by opening another or by pressing '$'),
the file is renamed, prompting the inotify-script to
run "notmuch new" and thereby update the notmuch tags.