feat(nvim): replace seanbreckenridge/yadm-git.vim with takinoy/yadm.nvim
I had weird problems where sometimes, the git repository path would
incorrectly change to the yadm repo path. For example, this started
to happen with nvim-orgmode/orgmode (but saw it previously already):

    Error executing vim.schedule lua callback: unhandled promise
    rejection: { "ENOENT: no such file or directory:
    fugitive:/Users/andreas/.local/share/yadm/repo.git/0/Documents/git_projects/notes/todos/sysadmin.org"
    } stack traceback:

    [C]: in function 'error'
    ...al/share/nvim/lazy/orgmode/lua/orgmode/utils/promise.lua:71: in
    function
    <...al/share/nvim/lazy/orgmode/lua/orgmode/utils/promise.lua:69>

I suppose it is some hard-to-debug race condition. Let's try this
plugin and see if it works better.